some minor fixes to the processing lambda
