+++ lambda/process_analytics/Cargo.toml
@@ -5,7 +5,6 @@ edition = "2021"
 
 [dependencies]
 aws_lambda_events = { version = "0.15.1", default-features = false, features = ["kinesis"] }
-base64 = "0.22.1"
 
 lambda_runtime = "0.13.0"
 tokio = { version = "1", features = ["macros"] }
